+++ sdk/communication/Azure.Communication.JobRouter/src/Models/RuleEngineQueueSelectorAttachment.cs
@@ -2,12 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Text.Json;
 using Azure.Core;
 
 namespace Azure.Communication.JobRouter
 {
-    public partial class RuleEngineQueueSelectorAttachment : IUtf8JsonSerializable
+    public partial class RuleEngineQueueSelectorAttachment
     {
         /// <summary> Initializes a new instance of RuleEngineQueueSelectorAttachment. </summary>
         /// <param name="rule">
@@ -32,16 +31,6 @@ namespace Azure.Communication.JobRouter
 
             Kind = QueueSelectorAttachmentKind.RuleEngine;
             Rule = rule;
-        }
-
-        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WritePropertyName("rule"u8);
-            writer.WriteObjectValue(Rule);
-            writer.WritePropertyName("kind"u8);
-            writer.WriteStringValue(Kind.ToString());
-            writer.WriteEndObject();
         }
     }
 }