+++ sdk/communication/Azure.Communication.JobRouter/src/Models/StaticQueueSelectorAttachment.cs
@@ -2,12 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Text.Json;
 using Azure.Core;
 
 namespace Azure.Communication.JobRouter
 {
-    public partial class StaticQueueSelectorAttachment : IUtf8JsonSerializable
+    public partial class StaticQueueSelectorAttachment
     {
         /// <summary> Initializes a new instance of StaticQueueSelectorAttachment. </summary>
         /// <param name="queueSelector">
@@ -21,16 +20,6 @@ namespace Azure.Communication.JobRouter
 
             Kind = QueueSelectorAttachmentKind.Static;
             QueueSelector = queueSelector;
-        }
-
-        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WritePropertyName("queueSelector"u8);
-            writer.WriteObjectValue(QueueSelector);
-            writer.WritePropertyName("kind"u8);
-            writer.WriteStringValue(Kind.ToString());
-            writer.WriteEndObject();
         }
     }
 }