+++ sdk/communication/Azure.Communication.Messages/tests/TemplateClient/MessageTemplateClientLiveTests.cs
@@ -2,9 +2,12 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Linq;
 using System.Threading.Tasks;
+using Azure.Communication.Messages.Models.Channels;
 using Azure.Core.TestFramework;
 using NUnit.Framework;
+using System.Collections.Generic;
 
 namespace Azure.Communication.Messages.Tests
 {
@@ -26,12 +29,13 @@ namespace Azure.Communication.Messages.Tests
 
             // Assert
             Assert.IsNotNull(templates);
-            var templatesEnumerable = templates.ToEnumerableAsync().Result;
+            List<MessageTemplateItem> templatesEnumerable = templates.ToEnumerableAsync().Result;
             Assert.IsNotEmpty(templatesEnumerable);
-            foreach (MessageTemplateItem template in templatesEnumerable)
+            foreach (WhatsAppMessageTemplateItem template in templatesEnumerable.Cast<WhatsAppMessageTemplateItem>())
             {
                 Assert.IsNotNull(template.Name);
                 Assert.IsNotNull(template.Language);
+                Assert.IsNotNull(template.Content);
             }
 
             return Task.CompletedTask;
@@ -49,12 +53,13 @@ namespace Azure.Communication.Messages.Tests
 
             // Assert
             Assert.IsNotNull(templates);
-            var templatesEnumerable = templates.ToEnumerableAsync().Result;
+            List<MessageTemplateItem> templatesEnumerable = templates.ToEnumerableAsync().Result;
             Assert.IsNotEmpty(templatesEnumerable);
-            foreach (MessageTemplateItem template in templatesEnumerable)
+            foreach (WhatsAppMessageTemplateItem template in templatesEnumerable.Cast<WhatsAppMessageTemplateItem>())
             {
                 Assert.IsNotNull(template.Name);
                 Assert.IsNotNull(template.Language);
+                Assert.IsNotNull(template.Content);
             }
 
             return Task.CompletedTask;