+++ sdk/core/Azure.Core/src/RequestUriBuilder.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Diagnostics;
+using System.IO;
 using System.Text;
 
 namespace Azure.Core
@@ -27,8 +28,6 @@ namespace Azure.Core
         private string? _host;
 
         private string? _scheme;
-
-        internal bool IsValidUri => Scheme != null && Host != null;
 
         /// <summary>
         /// Gets or sets the scheme name of the URI.