+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/CHANGELOG.md
@@ -1,14 +1,30 @@
 # Release History
 
-## 1.2.0-beta.1 (Unreleased)
+## 1.2.0-beta.1 (2024-02-09)
 
 ### Features Added
 
-### Breaking Changes
+* Added Azure Container Apps resource detector.
+  ([#41803](https://github.com/Azure/azure-sdk-for-net/pull/41803))
 
-### Bugs Fixed
+* Added `Azure.Monitor.OpenTelemetry.LiveMetrics`, enabling the sending of [live
+  metrics
+  data](https://learn.microsoft.com/azure/azure-monitor/app/live-stream).
+  The newly added `EnableLiveMetrics` property is set to `true` by default. This
+  property can be set to `false` to disable live metrics.
+  ([#41872](https://github.com/Azure/azure-sdk-for-net/pull/41872))
+
+- Added an experimental feature for logs emitted within an active tracing
+  context to follow the Activity's sampling decision. The feature can be enabled
+  by setting `OTEL_DOTNET_AZURE_MONITOR_EXPERIMENTAL_ENABLE_LOG_SAMPLING`
+  environment variable to `true`.
+  ([#41665](https://github.com/Azure/azure-sdk-for-net/pull/41665))
 
 ### Other Changes
+
+* Updated the vendored code in the `OpenTelemetry.ResourceDetectors.Azure`
+  resource detector to include the Azure Container Apps resource detector.
+  ([#41803](https://github.com/Azure/azure-sdk-for-net/pull/41803))
 
 ## 1.1.0 (2024-01-25)